добавил валидацию, создание/обноваление пользователя, enum-ы

+++ internal/models/user/user.go
@@ -9,7 +9,7 @@ type User struct {
 	Phone               sql.NullString `db:"phone"`
 	Password            sql.NullString `db:"password"`
 	Email               sql.NullString `db:"email"`
-	HoReCaId            sql.NullInt64  `db:"horeca_id"`
+	HoReCaId            sql.NullInt16  `db:"horeca_id"`
 	PasswordRecoveryUrl sql.NullString `db:"password_recovery_url"`
 	Messenger           sql.NullString `db:"messenger"`
 	CreatedAt           sql.NullTime   `db:"created_at"`
